model_signing/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
betterproto==2.0.0b6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
dnspython==2.6.1
email_validator==2.1.1
grpclib==0.4.7
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
id==1.4.0
idna==3.7
importlib-resources==5.13.0
in-toto-attestation==0.9.3
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
platformdirs==4.2.1
protobuf==5.26.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
requests==2.31.0
rfc8785==0.1.2
rich==13.7.1
securesystemslib==0.31.0
sigstore-protobuf-specs==0.3.1
sigstore-rekor-types==0.0.13
six==1.16.0
tuf==4.0.0
typing_extensions==4.11.0
urllib3==2.2.1
